+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - Remove ge-install-manager. The project has been archived.
 - Change Proton GE version from Proton-6.5-GE-2 to GE-Proton7-8.
 - Add PeaZip for an archive/compression utility.
+- Add BalenaEtcher as an image flashing utility.
 
 ## 2.2.0